Refactor main window into mixins
